moved functions calculating prescribed dosages out from component/reducer
- code clean up a bit
- Still need to move quantity calculation logic into components in PrescriptionForm
